### 2025-05-30 #245 @pibion
Wondering if energy-dependent convolution code is in scope for pyOpenSci
My team and I (especially @Arqu1100) have been working on energy-dependent convolutions for a nuclear physics application, and we are wondering if the code would be suitable for pyOpenSci review: https://github.com/det-lab/energyDependentColvolve.

We don't think this is suitable for JOSS because it is a single-function library.

However, the code is useful within the nuclear physics community and potentially the astrophysics/imaging community.  We ran into quite a few issues when writing the code and would like to help out other groups who need to do this "simple" calculation.  One of the strengths of this code base is the inclusion of test cases.  The only existing library we found that does what we do here is `varconvolve`, which we haven't been able to verify against our test cases.  Other examples of code that does a similar job are referenced on Stack Overflow as being deeply embedded into existing libraries (https://stackoverflow.com/questions/18624005/how-do-i-perform-a-convolution-in-python-with-a-variable-width-gaussian).
