### 2025-06-01 #245 @coatless
@pibion please resubmit following the pre-submission template guidelines.

https://github.com/pyOpenSci/software-submission/issues/new?template=presubmission-inquiry.md
