### 2025-06-02 #245 @jonas-eschle
@pibion did you reach out to domain-specific packages whether they are interested to incorporate the function? This seems a lot more promising to me
