### 2025-06-12 #245 @pibion
@jonas-eschle we've reached out to AstroPy, Numpy, and SciPy and it looks like the answer is no for each - this isn't quite in scope for AstroPy (and there is an existing package in their community that does this work!).  Numpy is very conservative about adding new features, and there wasn't any conclusive response from SciPy.

Would this package as a standalone be suitable for PyOpenSci or is it too small?
