### 2025-06-12 #245 @pibion
@coatless @jonas-eschle also opened a new issue: https://github.com/pyOpenSci/software-submission/issues/248
