### 2025-06-12 #245 @jonas-eschle
> [@jonas-eschle](https://github.com/jonas-eschle) we've reached out to AstroPy, Numpy, and SciPy and it looks like the answer is no for each - this isn't quite in scope for AstroPy (and there is an existing package in their community that does this work!).

Thanks for reaching out to them! When you say there is a package that does this, the immediate thought would be, what does this package (energyDependentConvolution) add to the community? What can it do that the other package can't?
(_please don't take this the wrong way that I am asking critically: software is easily built but hard to maintain. Buildinging is 20%, the rest that comes more like 80%, and if two packages do the same, that means time is lost which could be spent on one. So to make sure a package will survive, it also needs a niche to fit in and not duplicate existing functionality_)

> Numpy is very conservative about adding new features, and there wasn't any conclusive response from SciPy.

Could have guessed this one, but makes sense.

> Would this package as a standalone be suitable for PyOpenSci or is it too small?

I think if it is a standalone package (and fills a niche), there is nothing that would be too small, _as long as it is a standalone package_ and fulfills the requirements (setup, tests, docs, tutorials, etc.). The effort for a little package to set these things up is just proportionally higher.
